add whole month rendering and settings
for number of months to display at once. Make it possible to provide
templates for calendarHeader content or calendarWeekdays content.

The first day of the week should be changed via adding custom locales to
the required moment instance. As the locale data of moment doesn't include
information about which days are usually said to be weekends for a locale
there's a weekendDays setting that may be used for changing this.